# apps/backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Applad API
  version: 1.0.0
  description: Self-hosted Backend-as-a-Service API
  license:
    name: BSD-3-Clause
servers:
  - url: /v1
    description: Default (relative)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-Applad-Key
...
paths:
...
  /storage/buckets:
    get:
      summary: List buckets
      tags: [Storage]
      responses:
        '200':
          description: Bucket list
    post:
      summary: Create bucket
      tags: [Storage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                bucketId:
                  type: string
                name:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
                maximumFileSize:
                  type: integer
                allowedFileExtensions:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Bucket created
  /storage/buckets/{bucketId}/files:
    parameters:
      - name: bucketId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: List files
      tags: [Storage]
      responses:
        '200':
          description: File list
    post:
      summary: Upload file
      tags: [Storage]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                fileId:
                  type: string
                file:
                  type: string
                  format: binary
                permissions:
                  type: string
      responses:
        '201':
          description: File uploaded
...